+++ packages/validate-openalex/src/actions/actors.ts
@@ -1,22 +1,28 @@
 import { Effect } from 'effect';
-import { getORCID } from '../context';
+import { getContext, getORCID } from '../context';
 import { setEventsStore } from '../store/setter';
 import { getAuthorAlternativeStrings } from './tester';
-import { text, select, events2options } from '../prompt';
+import { text, select, events2options, confirm } from '../prompt';
 import { asORCID } from '@univ-lehavre/biblio-openalex-types';
-import { searchAuthorByName, searchAuthorByORCID } from '../fetch';
+import { searchAuthorByName, searchAuthorByORCID, searchWorksByAuthorIDs } from '../fetch';
 import { ContextStore, EventsStore, updateContextStore, updateEventsStore } from '../store';
 import {
   buildAuthorResultsPendingEvents,
+  buildEvent,
+  buildReference,
   getEvents,
   getManyEvent,
+  getOpenAlexIDs,
+  getStatusOfAffiliation,
+  getStatusOfAuthorDisplayNameAlternative,
   isInteresting,
   removeDuplicates,
   updateNewEventsWithExistingMetadata,
 } from '../events';
-import type { AuthorsResult, ORCID } from '@univ-lehavre/biblio-openalex-types';
+import type { AuthorsResult, ORCID, WorksResult } from '@univ-lehavre/biblio-openalex-types';
 import type { ConfigError } from 'effect/ConfigError';
-import type { IEvent } from '../events/types';
+import type { IEvent, Status } from '../events/types';
+import { IContext } from '../context/types';
 
 const insert_new_ORCID = (): Effect.Effect<void, Error | ConfigError, ContextStore | EventsStore> =>
   Effect.gen(function* () {
@@ -97,7 +103,210 @@ const removeAuthorPendings = (): Effect.Effect<void, Error, ContextStore | Event
     yield* setEventsStore(notPendings);
   });
 
-const extendsToWorks = () => Effect.gen(function* () {});
+const checkWork = (orcid: ORCID, authorOpenalexID: string, work: WorksResult) =>
+  Effect.gen(function* () {
+    // On regarde chaque publication
+    const authorships = work.authorships;
+    if (authorships.length === 0) return;
+    const authorship = authorships.find(author => author.author.id === authorOpenalexID);
+    if (authorship === undefined) {
+      const event = yield* buildEvent({
+        from: work.id,
+        id: orcid,
+        entity: 'work',
+        field: 'id',
+        value: work.id,
+        label: buildReference(work),
+        status: 'rejected',
+      });
+      yield* updateEventsStore([event]);
+      return;
+    }
+    const status = getStatusOfAuthorDisplayNameAlternative(
+      authorship.raw_author_name,
+      orcid,
+      yield* getEvents(),
+    );
+    if (status === 'rejected') {
+      const event = yield* buildEvent({
+        from: work.id,
+        id: orcid,
+        entity: 'work',
+        field: 'id',
+        value: work.id,
+        label: buildReference(work),
+        status: 'rejected',
+      });
+      yield* updateEventsStore([event]);
+      return;
+    } else if (status === undefined) {
+      const confirmed = yield* confirm(
+        `Acceptez-vous d'ajouter "${authorship.raw_author_name}" comme forme imprimée pour l'ORCID ${orcid} ?`,
+      );
+      if (typeof confirmed !== 'boolean') throw new Error('Réponse invalide');
+      if (confirmed) {
+        const event = yield* buildEvent({
+          from: work.id,
+          id: orcid,
+          entity: 'author',
+          field: 'display_name_alternatives',
+          value: authorship.raw_author_name,
+          status: 'accepted',
+        });
+        yield* updateEventsStore([event]);
+      } else {
+        const event = yield* buildEvent({
+          from: work.id,
+          id: orcid,
+          entity: 'author',
+          field: 'display_name_alternatives',
+          value: authorship.raw_author_name,
+          status: 'rejected',
+        });
+        const event2 = yield* buildEvent({
+          from: work.id,
+          id: orcid,
+          entity: 'work',
+          field: 'id',
+          value: work.id,
+          label: buildReference(work),
+          status: 'rejected',
+        });
+        yield* updateEventsStore([event, event2]);
+        return;
+      }
+    }
+    // On passe aux affiliations
+    for (const affiliation of authorship.affiliations) {
+      const raw_affiliation_string = affiliation.raw_affiliation_string;
+      // On regarde s’il n’y a pas déjà une affiliation rejetée
+      for (const institutionID of affiliation.institution_ids) {
+        const status: Status | undefined = getStatusOfAffiliation(
+          institutionID,
+          orcid,
+          yield* getEvents(),
+        );
+        if (status === 'rejected') {
+          const toPush: IEvent[] = [];
+          toPush.push(
+            yield* buildEvent({
+              from: work.id,
+              id: orcid,
+              entity: 'work',
+              field: 'id',
+              value: work.id,
+              label: buildReference(work),
+              status: 'rejected',
+            }),
+          );
+          toPush.push(
+            yield* buildEvent({
+              from: work.id,
+              id: orcid, // Irrelevant here
+              entity: 'institution',
+              field: 'display_name_alternatives',
+              value: raw_affiliation_string,
+              status: 'rejected',
+            }),
+          );
+          yield* updateEventsStore(toPush);
+          return;
+        } else if (status === undefined) {
+          const selected = yield* confirm(
+            `${raw_affiliation_string}: Est-ce une affiliation valide pour ce chercheur ?`,
+          );
+          if (typeof selected !== 'boolean') throw new Error('Réponse invalide');
+          if (selected) {
+            const toPush: IEvent[] = [];
+            toPush.push(
+              yield* buildEvent({
+                from: work.id,
+                id: orcid, // Irrelevant here
+                entity: 'institution',
+                field: 'display_name_alternatives',
+                value: raw_affiliation_string,
+                status: 'accepted',
+              }),
+            );
+            yield* updateEventsStore(toPush);
+          } else {
+            const toPush: IEvent[] = [];
+            toPush.push(
+              yield* buildEvent({
+                from: work.id,
+                id: orcid, // Irrelevant here
+                entity: 'institution',
+                field: 'display_name_alternatives',
+                value: raw_affiliation_string,
+                status: 'rejected',
+              }),
+            );
+            toPush.push(
+              yield* buildEvent({
+                from: work.id,
+                id: orcid,
+                entity: 'work',
+                field: 'id',
+                value: work.id,
+                label: buildReference(work),
+                status: 'rejected',
+              }),
+            );
+            for (const institutionID of affiliation.institution_ids) {
+              toPush.push(
+                yield* buildEvent({
+                  from: work.id,
+                  id: orcid,
+                  entity: 'author',
+                  field: 'affiliation',
+                  value: institutionID,
+                  status: 'rejected',
+                }),
+              );
+            }
+            yield* updateEventsStore(toPush);
+            return;
+          }
+        }
+      }
+      // Si elle est inconnue, on demande à l’utilisateur
+      // si il confirme, on émet un événement d’acceptation pour la forme imprimée de l’affiliation
+      // si il refuse, on émet un événement de rejet pour la publication et de la forme imprimée de l’affiliation
+    }
+    const event: IEvent = yield* buildEvent({
+      from: work.id,
+      id: orcid,
+      entity: 'work',
+      field: 'id',
+      value: work.id,
+      label: buildReference(work),
+      status: 'accepted',
+    });
+    const event2: IEvent = yield* buildEvent({
+      from: work.id,
+      id: orcid,
+      entity: 'author',
+      field: 'openalexID',
+      value: authorOpenalexID,
+      status: 'accepted',
+    });
+    yield* updateEventsStore([event, event2]);
+  });
+
+const extendsToWorks = () =>
+  Effect.gen(function* () {
+    const { id }: IContext = yield* getContext();
+    if (id === undefined) return;
+    const openalexIDs = getOpenAlexIDs(id, yield* getEvents());
+    for (const authorOpenalexID of openalexIDs) {
+      // On travaille sur un chercheur
+      const works = yield* searchWorksByAuthorIDs([authorOpenalexID]);
+      if (works.length === 0) continue;
+      for (const work of works) {
+        yield* checkWork(id, authorOpenalexID, work);
+      }
+    }
+  });
 
 export {
   insert_new_ORCID,